+++ src/defaults.js
@@ -1,41 +1,5 @@
 import Handlebars from 'handlebars';
-import yaml from 'js-yaml';
-import marked from 'marked';
-import frontMatter from 'front-matter';
-
-var parsers = {
-  content: {
-    pattern: '\.(html|hbs|handlebars)$',
-    parseFn: (contents, path) => {
-      var matter = frontMatter(contents);
-      return {
-        contents: matter.body,
-        data: matter.attributes
-      };
-    }
-  },
-  markdown: {
-    pattern: '\.(md|markdown)$',
-    parseFn: (contents, path) => {
-      var matter = frontMatter(contents);
-      return {
-        contents: marked(matter.body),
-        data: matter.attributes
-      };
-    }
-  },
-  yaml: {
-    pattern: '\.(yaml|yml)$',
-    parseFn: (contents, path) => ({ contents: yaml.safeLoad(contents) })
-  },
-  json: {
-    pattern: '\.json$',
-    parseFn: (contents, path) => ({ contents: JSON.parse(contents) })
-  },
-  default: {
-    parseFn: (contents, path) => ({ contents: contents })
-  }
-};
+import parsers from './parse/parsers';
 
 const defaults = {
   src: {
