+++ src/parse/parsers.js
@@ -1,0 +1,43 @@
+/**
+ * Default parsers, merged into options. Can be overridden by user options
+ */
+
+import yaml from 'js-yaml';
+import marked from 'marked';
+import frontMatter from 'front-matter';
+
+var parsers = {
+  content: {
+    pattern: '\.(html|hbs|handlebars)$',
+    parseFn: (contents, path) => {
+      var matter = frontMatter(contents);
+      return {
+        contents: matter.body,
+        data: matter.attributes
+      };
+    }
+  },
+  markdown: {
+    pattern: '\.(md|markdown)$',
+    parseFn: (contents, path) => {
+      var matter = frontMatter(contents);
+      return {
+        contents: marked(matter.body),
+        data: matter.attributes
+      };
+    }
+  },
+  yaml: {
+    pattern: '\.(yaml|yml)$',
+    parseFn: (contents, path) => ({ contents: yaml.safeLoad(contents) })
+  },
+  json: {
+    pattern: '\.json$',
+    parseFn: (contents, path) => ({ contents: JSON.parse(contents) })
+  },
+  default: {
+    parseFn: (contents, path) => ({ contents: contents })
+  }
+};
+
+export default parsers;
